Aufstellung: aktuellen Platz anzeigen, Speichern-Knopf entfernen
Ganz links in der Spielerliste steht jetzt der Platz, auf dem der Spieler
gerade aufgestellt ist - Feld-Slot, Bankplatz oder ein Strich. Aufgestellte
Spieler stehen in Akzentfarbe, damit auf einen Blick erkennbar ist, wer noch
frei ist. Die Spalte wird an allen sechs Stellen nachgezogen, an denen sich
die Aufstellung aendern kann; dabei wird nur die betroffene Zelle beschrieben,
sodass die Scrollposition erhalten bleibt.

Der Speichern-Knopf entfaellt - jede Aenderung wird ohnehin sofort
uebernommen. Dabei fiel auf, dass ein Formationswechsel bisher nie
automatisch gespeichert wurde; der Knopf war der einzige Weg. Ohne ihn waere
das ein Datenverlust geworden, deshalb speichert jetzt auch der
Formationswechsel sofort. Rueckmeldung gibt es nur noch im Fehlerfall.

Die Gesamtstaerke wird mit einer Nachkommastelle angezeigt.

+++ scripts/models/AufstellungModel.cs
@@ -8,7 +8,10 @@ public class AufstellungModel
     public long VereinId { get; set; }
     public string Formation { get; set; } = "4-4-2";
     public Dictionary<string, long> Positionen { get; set; } = new();
-    public int Gesamtstaerke { get; set; }
+    /// <summary>
+    /// Durchschnittliche Stärke über alle Positionen der Formation – unbesetzte zählen 0.
+    /// </summary>
+    public double Gesamtstaerke { get; set; }
 
     /// <summary>Ersatzbank in Reihenfolge (Spieler-IDs).</summary>
     public List<long> Ersatzbank { get; set; } = new();
